2CentsCapitalTask/requirements.txt
python-docx==0.8.11
gradio==4.44.0
streamlit==1.28.1
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.0
tiktoken==0.5.2
faiss-cpu==1.7.4
PyMuPDF==1.23.8
spacy==3.7.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
